director/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aiofiles==0.6.0
    # via quart
aiohttp==3.7.4.post0
    # via
    #   twitchio
    #   wled
appdirs==1.4.4
    # via black
aspy.refactor-imports==2.2.0
    # via reorder-python-imports
async-timeout==3.0.1
    # via
    #   aiohttp
    #   twitchio
attrs==20.3.0
    # via
    #   aiohttp
    #   pytest
backoff==1.10.0
    # via wled
black==20.8b1
    # via -r requirements.in
blinker==1.4
    # via quart
cached-property==1.5.2
    # via aspy.refactor-imports
chardet==4.0.0
    # via aiohttp
click==7.1.2
    # via
    #   black
    #   quart
evdev==1.4.0
    # via -r requirements.in
h11==0.12.0
    # via
    #   hypercorn
    #   wsproto
h2==4.0.0
    # via hypercorn
hpack==4.0.0
    # via h2
hypercorn==0.11.2
    # via
    #   -r requirements.in
    #   quart
hyperframe==6.0.0
    # via h2
idna==3.1
    # via yarl
iniconfig==1.1.1
    # via pytest
itsdangerous==1.1.0
    # via quart
jinja2==2.11.3
    # via quart
markupsafe==1.1.1
    # via jinja2
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via black
obs-websocket-py==0.5.3
    # via -r requirements.in
packaging==20.9
    # via
    #   pytest
    #   wled
pathspec==0.8.1
    # via black
pluggy==0.13.1
    # via pytest
priority==1.3.0
    # via hypercorn
psutil==5.8.0
    # via -r requirements.in
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest==6.2.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==0.17.0
    # via -r requirements.in
python-slugify==5.0.2
    # via -r requirements.in
pytz==2021.3
    # via -r requirements.in
pyyaml==5.4.1
    # via -r requirements.in
quart==0.14.1
    # via -r requirements.in
regex==2021.3.17
    # via black
reorder-python-imports==2.4.0
    # via -r requirements.in
six==1.15.0
    # via
    #   obs-websocket-py
    #   python-dateutil
    #   websocket-client
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   black
    #   hypercorn
    #   pytest
    #   quart
twitchio==1.2.2
    # via -r requirements.in
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   black
websocket-client==0.58.0
    # via obs-websocket-py
websockets==8.1
    # via twitchio
werkzeug==1.0.1
    # via quart
wled==0.4.4
    # via -r requirements.in
wsproto==1.0.0
    # via hypercorn
yarl==1.6.3
    # via
    #   aiohttp
    #   wled
